+++ oemof_flexmex/model_structure/component_attrs/electricity-curtailment.csv
