+++ oemof_flexmex/model_structure/component_attrs/electricity-liion_battery.csv
