+++ oemof_flexmex/model_structure/component_attrs/uranium-nuclear-st.csv
